+++ packages/ai-constructs/src/conversation/conversation_handler_construct.ts
@@ -35,6 +35,12 @@ export type ConversationHandlerFunctionProps = {
     region?: string;
   }>;
   /**
+   * An amount of memory (RAM) to allocate to the function between 128 and 10240 MB.
+   * Must be a whole number.
+   * Default is 512MB.
+   */
+  memoryMB?: number;
+  /**
    * @internal
    */
   outputStorageStrategy?: BackendOutputStorageStrategy<AIConversationOutput>;
@@ -86,6 +92,7 @@ export class ConversationHandlerFunction
         timeout: Duration.seconds(60),
         entry: this.props.entry ?? defaultHandlerFilePath,
         handler: 'handler',
+        memorySize: this.resolveMemory(),
         bundling: {
           // Do not bundle SDK if conversation handler is using our default implementation which is
           // compatible with Lambda provided SDK.
@@ -117,7 +124,10 @@ export class ConversationHandlerFunction
       conversationHandler.addToRolePolicy(
         new PolicyStatement({
           effect: Effect.ALLOW,
-          actions: ['bedrock:InvokeModel'],
+          actions: [
+            'bedrock:InvokeModel',
+            'bedrock:InvokeModelWithResponseStream',
+          ],
           resources,
         })
       );
@@ -150,4 +160,27 @@ export class ConversationHandlerFunction
       },
     });
   };
+
+  private resolveMemory = () => {
+    const memoryMin = 128;
+    const memoryMax = 10240;
+    const memoryDefault = 512;
+    if (this.props.memoryMB === undefined) {
+      return memoryDefault;
+    }
+    if (
+      !isWholeNumberBetweenInclusive(this.props.memoryMB, memoryMin, memoryMax)
+    ) {
+      throw new Error(
+        `memoryMB must be a whole number between ${memoryMin} and ${memoryMax} inclusive`
+      );
+    }
+    return this.props.memoryMB;
+  };
 }
+
+const isWholeNumberBetweenInclusive = (
+  test: number,
+  min: number,
+  max: number
+) => min <= test && test <= max && test % 1 === 0;