+++ packages/integration-tests/src/test-e2e/deployment/reference_auth_project.deployment.test.ts
@@ -1,0 +1,4 @@
+import { ReferenceAuthTestProjectCreator } from '../../test-project-setup/reference_auth_project.js';
+import { defineDeploymentTest } from './deployment.test.template.js';
+
+defineDeploymentTest(new ReferenceAuthTestProjectCreator());