+++ packages/integration-tests/src/test-e2e/sandbox/data_storage_auth_with_triggers.sandbox.test.ts
@@ -1,0 +1,4 @@
+import { defineSandboxTest } from './sandbox.test.template.js';
+import { DataStorageAuthWithTriggerTestProjectCreator } from '../../test-project-setup/data_storage_auth_with_triggers.js';
+
+defineSandboxTest(new DataStorageAuthWithTriggerTestProjectCreator());