+++ packages/integration-tests/src/test-projects/conversation-handler/amplify/constants.ts
@@ -11,7 +11,6 @@ export const bedrockModelId = 'anthropic.claude-3-haiku-20240307-v1:0';
 export const expectedTemperaturesInProgrammaticToolScenario = {
   Seattle: 75,
   Boston: 58,
-  Miami: 97,
 };
 
 export const expectedTemperatureInDataToolScenario = 85;