+++ packages/backend-function/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @aws-amplify/backend-function
 
+## 1.7.2
+
+### Patch Changes
+
+- 601a2c1: dedupe environment variables in amplify env type generator
+
+## 1.7.1
+
+### Patch Changes
+
+- bd4ff4d: Fix jsdocs that incorrectly state default memory settings
+- Updated dependencies [5f46d8d]
+  - @aws-amplify/backend-output-schemas@1.4.0
+
 ## 1.7.0
 
 ### Minor Changes